Added kernel module
The kernel module for drivers/misc is added as well as patch for linux-3.4-rc4
which adds compilation of memnotify as module by default.

Due to this module did not accepted in MM even published and discussed it has
sense to share this portable and simple code together with library.

Signed-off-by: Leonid Moiseichuk <leonid.moiseichuk@nokia.com>